🚧Convert to structs (#83)
* Start work for issue #80

* ide: add perf results as a solution folder

* test: add engine perf project to the solution

* ci: add perf project to build status check workflow

* docs: fix grammary issues in comments

* chore!: convert particle to readonly record struct

* chore: move particle update processing and logic to particle pool

* refactor: make simple improvements

* fix: add content load state checks to content methods

* test: remove particle tests

* refactor: make simple improvements to array creation

* docs: improve comments

* test: set up simple perf tests

* docs: update perf results

* docs: fix grammar issues in comments

* refactor: add simple code improvements

* perf: improve perf project

* perf: implement 0.62% to 5.15% perf improvement

This is based on how many particles are being processed

* refactor!: change property setter scopes

* refactor: convert prop to auto prop

* test: improve performance tests

* perf: collect perf results

* refactor: fix grammar issues, improve code for standards, and clean up code

* perf: performance improvement

asdf

* refactor: improve code readability

* ide: adjust project run configs

* test: create shared project for perf projects to utilize

* test: create perf project for particle pool

* docs: update and add perf results

* chore: add kill all particles on click to color scene

* docs: improve grammar/spelling in code docs

* refactor!: all rate values types changed from int to floating point

* refactor!: changed the particle effect class to readonly record struct

* chore: updated testing scenes to accommodate changes

* cleanup: improve code in perf test project

* chore!: changed easing function types from double to float

* chore!: changed behavior props from double to float

* chore!: changed return type of easing random behavior settings props

* chore!: changed type for easing random behavior props

* chore: update color scene literal values to floats

* chore!: moved the add and remove behavior logic to the particle pool

* chore: add default constructor to ensure that prop is init

* chore: perform null checks on behavior property

* chore: add perf project improvements

* test: fix and improve tests

* test,perf: improve perf project set up

* refactor: change foreach loops to for loops

* config: set internals visible to unit test project

* chore!: add disposal checks to class members

* test,chore: improve content loading and unloading in scenes

* chore: change return type of factory to interface

* refactor: remove unneccesary code

* test: increase code coverage for particle pool class

* ide: ignore code coverage for perf projects

* test: increase code coverage of particle struct

* test: increase code coverage of particle engine class

* test: increase code coverage of particle effect struct

* test: increase code coverage of extension methods

* test: increase code coverage of easing random behavior

* chore: add settings property to easing random behaviors

* chore: setup default values for default constructor

+++ Plazma/Behaviors/EasingRandomBehaviorSettings.cs
@@ -50,7 +50,7 @@ public readonly record struct EasingRandomBehaviorSettings
     /// </para>
     /// </remarks>
     [JsonIgnore]
-    public Func<double, double>? UpdateValue { get; init; }
+    public Func<float, float>? UpdateValue { get; init; }
 
     /// <summary>
     /// Gets a delegate that will give the current value of the behavior and
@@ -65,7 +65,7 @@ public readonly record struct EasingRandomBehaviorSettings
     /// </para>
     /// </remarks>
     [JsonIgnore]
-    public Func<double, float>? UpdateRandomStartMin { get; init; }
+    public Func<float, float>? UpdateRandomStartMin { get; init; }
 
     /// <summary>
     /// Gets a delegate that will give the current value of the behavior and
@@ -80,7 +80,7 @@ public readonly record struct EasingRandomBehaviorSettings
     /// </para>
     /// </remarks>
     [JsonIgnore]
-    public Func<double, float>? UpdateRandomStartMax { get; init; }
+    public Func<float, float>? UpdateRandomStartMax { get; init; }
 
     /// <summary>
     /// Gets the minimum amount of change used in randomization.
@@ -101,12 +101,12 @@ public readonly record struct EasingRandomBehaviorSettings
     /// <summary>
     /// Gets the minimum total amount of time in milliseconds to complete the behavior.
     /// </summary>
-    /// <remarks>A value less then or equal to 0 will result in the behavior not working.</remarks>
+    /// <remarks>A value less than or equal to 0 will result in the behavior not working.</remarks>
     public float LifeTimeMillisecondsMin { get; init; }
 
     /// <summary>
     /// Gets the maximum total amount of time in milliseconds to complete the behavior.
     /// </summary>
-    /// <remarks>A value less then or equal to 0 will result in the behavior not working.</remarks>
+    /// <remarks>A value less than or equal to 0 will result in the behavior not working.</remarks>
     public float LifeTimeMillisecondsMax { get; init; }
 }
