+++ Plazma/Behaviors/IBehavior.cs
@@ -14,12 +14,12 @@ public interface IBehavior
     /// <summary>
     /// Gets the current value of the behavior.
     /// </summary>
-    double Value { get; }
+    float Value { get; }
 
     /// <summary>
     /// Gets the current amount of time that has elapsed for the behavior.
     /// </summary>
-    double ElapsedTime { get; }
+    float ElapsedTime { get; }
 
     /// <summary>
     /// Gets the particle attribute to apply the behavior result to.
@@ -32,13 +32,13 @@ public interface IBehavior
     bool Enabled { get; }
 
     /// <summary>
-    /// Gets the life time of the behavior in milliseconds.
+    /// Gets the lifetime of the behavior in milliseconds.
     /// </summary>
     /// <remarks>
-    ///     Once the amount of time has elapsed the life time of the
+    ///     Once the amount of time has elapsed the lifetime of the
     ///     behavior, the behavior will be disabled.
     /// </remarks>
-    double LifeTime { get; }
+    float LifeTime { get; }
 
     /// <summary>
     /// Updates the behavior.
